Quick fix for multiple sticky posts
Upon stickying two posts on my subreddit, I realised that there was a gap between them.
Imagine it so:
_________________________
_________________________

It was horrendous. But, to my relief and thanks, all I had to do was simply remove
`margin-bottom: 8px` on `.link` (or reset it, at least).
This reset only applies to stickied links.
I also properly updated the version number :)