Remove requirement for single quotes for `promoted_tile`
`promoted_tile` is a URL that matches a user-selected link in the
sidebar, for them to promote something. It used to require double and
single quotes, “‘like this’”.

Instead, some clever Stylus stuff really let me have this fixed all
along. Now a user can just write “https://github.com/huw/slight” and be
on their way!

Also, because this is the first in a series of features that I’ll be
adding over the next few days, this is now Slight v1.1 😄