Fix misaligned doman in link
For a while, `domain` has been difficult to properly align with the
rest of the stuff—after all, we’re pushing it below its normal position
using only CSS.

Recent reddit changes messed this up, so a new solution was needed.
This one sets the vertical align to `text-top`, which is then rectified
to put the domain name at the same position on all devices. Awesome!