Fix tagline misalignment with image flairs
Users on /r/Monstercat reported problems with their taglines being
misaligned on some browsers. I, a Safari user, had not noticed this,
but Blink-based renderers (and possible more) appeared to be breaking
them.

Upon arriving at a successful fix, I found that it broke if text-based
flairs were enabled (which the majority of Slight subreddits use). It
was easy, then, to only enable the changes if the subreddit was using
image flairs. Let’s hope this works.