Fix RES comment boxes with dark body
When `light_body = false`, the RES comment boxes would still show up in white. Because we didn't want to have extra, cloggy code, we didn't bother changing these from the default colours which RES provided.

The fix was simple: Give these a colour in dark mode. There was pre-existing Naut code, which was just adapted to the circumstances for us. Easy.